Fitting now works completely on TransitionSeries, rather than on Elements. Arbitrary pile-up and summation peaks can now be added. TransitionSeries now serializable for the purpose of saving settings.

+++ Peakaboo_DataTypes/src/peakaboo/datatypes/Ratios.java
@@ -13,7 +13,7 @@ public class Ratios
 	
 	public static String fromFloat(float value, boolean integersOnly)
 	{
-		double ratioValue = Math.pow(logValue, Math.abs(value));
+		float ratioValue = (float)Math.pow(logValue, Math.abs(value));
 		int decimals = 0;;
 		if (ratioValue < logValue && !integersOnly) decimals = 1;
